refactor(sales): reorganize deal product module and improve bulk product select
Split useDeals.tsx and product hooks/components into focused files under
mutations/, payment/, product-table/, and product-list/ subfolders, add a
ProductEditSheet for swapping deal line products, and extend the shared
SelectProductsBulk (ui-modules) with controlled open state, selection
limits, and custom labels so core-ui and accounting_ui can reuse it.
