+++ frontend/libs/erxes-ui/src/components/toaster.tsx
@@ -2,13 +2,29 @@ import { Toast } from './toasts';
 import { useToast } from '../hooks/use-toast';
 
 export function Toaster() {
-  const { toasts } = useToast();
+  const { toasts, dismiss } = useToast();
 
   return (
     <Toast.Provider>
-      {toasts.map(function ({ id, title, description, action, ...props }) {
+      {toasts.map(function ({
+        id,
+        title,
+        description,
+        action,
+        dismissOnPointerEnter,
+        ...props
+      }) {
         return (
-          <Toast key={id} {...props}>
+          <Toast
+            key={id}
+            {...props}
+            onPointerEnter={(event) => {
+              props.onPointerEnter?.(event);
+              if (dismissOnPointerEnter) {
+                dismiss(id);
+              }
+            }}
+          >
             <div className="grid gap-1">
               {title && <Toast.Title>{title}</Toast.Title>}
               {description && (
