+++ frontend/libs/ui-modules/src/modules/products/components/SelectProductsBulk.tsx
@@ -21,9 +21,37 @@ import { SelectCompany } from '../../contacts/components/SelectCompany';
 import { SelectCategory } from '../categories/components/SelectCategory';
 
 interface SelectProductsProps {
-  onSelect: (productIds: string[], products?: IProduct[]) => void;
-  children: React.ReactNode;
+  onSelect: (
+    productIds: string[],
+    products?: IProduct[],
+  ) => void | Promise<void>;
+  children?: React.ReactNode;
   productIds?: string[];
+  initialProducts?: IProduct[];
+  open?: boolean;
+  onOpenChange?: (open: boolean) => void;
+  selectionLimit?: number;
+  title?: React.ReactNode;
+  description?: React.ReactNode;
+  submitLabel?: React.ReactNode;
+  cancelLabel?: React.ReactNode;
+  selectedLabel?: React.ReactNode;
+  isSelectionValid?: (productIds: string[], products: IProduct[]) => boolean;
+}
+
+interface SelectProductsBulkContentProps
+  extends Pick<
+      SelectProductsProps,
+      | 'onSelect'
+      | 'productIds'
+      | 'initialProducts'
+      | 'selectionLimit'
+      | 'isSelectionValid'
+    >,
+    Required<
+      Pick<SelectProductsProps, 'submitLabel' | 'cancelLabel' | 'selectedLabel'>
+    > {
+  setOpen: (open: boolean) => void;
 }
 
 interface ProductsListProps {
@@ -31,30 +59,60 @@ interface ProductsListProps {
   setSelectedProducts: React.Dispatch<React.SetStateAction<IProduct[]>>;
   selectedProductIds: string[];
   setSelectedProductIds: React.Dispatch<React.SetStateAction<string[]>>;
+  selectionLimit?: number;
 }
 
 export const SelectProductsBulk = ({
   onSelect,
   children,
   productIds,
+  initialProducts,
+  open,
+  onOpenChange,
+  selectionLimit,
+  title = 'Select Products',
+  description,
+  submitLabel = 'Add Many Products',
+  cancelLabel = 'Cancel',
+  selectedLabel = 'Added',
+  isSelectionValid,
 }: SelectProductsProps) => {
-  const [open, setOpen] = useState(false);
+  const [internalOpen, setInternalOpen] = useState(false);
+  const resolvedOpen = open ?? internalOpen;
+
+  const setOpen = (nextOpen: boolean) => {
+    if (open === undefined) {
+      setInternalOpen(nextOpen);
+    }
+    onOpenChange?.(nextOpen);
+  };
 
   return (
-    <Sheet open={open} onOpenChange={setOpen}>
-      <Sheet.Trigger asChild>{children}</Sheet.Trigger>
+    <Sheet open={resolvedOpen} onOpenChange={setOpen}>
+      {children && <Sheet.Trigger asChild>{children}</Sheet.Trigger>}
       <Sheet.View className="sm:max-w-6xl">
         <Sheet.Header>
           <div>
-            <Sheet.Title>Select Products</Sheet.Title>
+            <Sheet.Title>{title}</Sheet.Title>
+            {description && (
+              <Sheet.Description>{description}</Sheet.Description>
+            )}
           </div>
           <Sheet.Close />
         </Sheet.Header>
-        <SelectProductsBulkContent
-          setOpen={setOpen}
-          onSelect={onSelect}
-          productIds={productIds}
-        />
+        {resolvedOpen && (
+          <SelectProductsBulkContent
+            setOpen={setOpen}
+            onSelect={onSelect}
+            productIds={productIds}
+            initialProducts={initialProducts}
+            selectionLimit={selectionLimit}
+            submitLabel={submitLabel}
+            cancelLabel={cancelLabel}
+            selectedLabel={selectedLabel}
+            isSelectionValid={isSelectionValid}
+          />
+        )}
       </Sheet.View>
     </Sheet>
   );
@@ -64,31 +122,49 @@ const SelectProductsBulkContent = ({
   setOpen,
   onSelect,
   productIds,
-}: {
-  setOpen: (open: boolean) => void;
-  onSelect: (productIds: string[], products?: IProduct[]) => void;
-  productIds?: string[];
-}) => {
-  const [selectedProductIds, setSelectedProductIds] = useState<string[]>([]);
-  const [selectedProducts, setSelectedProducts] = useState<IProduct[]>([]);
-
-  useEffect(() => {
-    if (productIds?.length) {
-      setSelectedProductIds(productIds);
-    }
-  }, [productIds]);
+  initialProducts,
+  selectionLimit,
+  submitLabel,
+  cancelLabel,
+  selectedLabel,
+  isSelectionValid,
+}: SelectProductsBulkContentProps) => {
+  const [selectedProductIds, setSelectedProductIds] = useState<string[]>(
+    productIds || [],
+  );
+  const [selectedProducts, setSelectedProducts] = useState<IProduct[]>(
+    initialProducts || [],
+  );
+  const [submitting, setSubmitting] = useState(false);
 
   const handleAddProduct = (data: { productsAdd: { _id: string } }) => {
+    if (
+      selectionLimit !== undefined &&
+      selectedProductIds.length >= selectionLimit
+    ) {
+      return;
+    }
     setSelectedProductIds((prev) => [...prev, data.productsAdd._id]);
   };
 
-  const handleSelect = () => {
-    onSelect(
-      selectedProducts.map((p) => p._id),
-      selectedProducts,
-    );
-    setOpen(false);
+  const handleSelect = async () => {
+    setSubmitting(true);
+    try {
+      await onSelect(selectedProductIds, selectedProducts);
+      setOpen(false);
+    } catch {
+      // The consuming mutation displays its own error state.
+    } finally {
+      setSubmitting(false);
+    }
   };
+  const allSelectedProductsResolved = selectedProductIds.every((productId) =>
+    selectedProducts.some((product) => product._id === productId),
+  );
+  const selectionLimitReached =
+    selectionLimit !== undefined && selectedProductIds.length >= selectionLimit;
+  const selectionIsValid =
+    isSelectionValid?.(selectedProductIds, selectedProducts) ?? true;
 
   return (
     <>
@@ -98,12 +174,14 @@ const SelectProductsBulkContent = ({
           selectedProductIds={selectedProductIds}
           setSelectedProductIds={setSelectedProductIds}
           setSelectedProducts={setSelectedProducts}
+          selectionLimit={selectionLimit}
         />
         <SelectedProductsList
           selectedProducts={selectedProducts}
           selectedProductIds={selectedProductIds}
           setSelectedProductIds={setSelectedProductIds}
           setSelectedProducts={setSelectedProducts}
+          selectedLabel={selectedLabel}
         />
       </Sheet.Content>
       <Sheet.Footer className="sm:justify-between">
@@ -112,14 +190,29 @@ const SelectProductsBulkContent = ({
             onCompleted: handleAddProduct,
             refetchQueries: [GET_PRODUCTS],
           }}
-        />
-        <div className="flex gap-2 items-center">
+        >
+          <Button variant="outline" disabled={selectionLimitReached}>
+            <IconPlus />
+            Create new product
+          </Button>
+        </AddProduct>
+        <div className="ml-auto flex gap-2 items-center">
           <Sheet.Close asChild>
             <Button variant="secondary" className="bg-border">
-              Cancel
+              {cancelLabel}
             </Button>
           </Sheet.Close>
-          <Button onClick={handleSelect}>Add Many Products</Button>
+          <Button
+            onClick={handleSelect}
+            disabled={
+              submitting ||
+              selectedProductIds.length === 0 ||
+              !allSelectedProductsResolved ||
+              !selectionIsValid
+            }
+          >
+            {submitLabel}
+          </Button>
         </div>
       </Sheet.Footer>
     </>
@@ -130,6 +223,7 @@ const ProductsList = ({
   setSelectedProducts,
   selectedProductIds,
   setSelectedProductIds,
+  selectionLimit,
 }: ProductsListProps) => {
   const [search, setSearch] = useState(
     () => localStorage.getItem('search') || '',
@@ -162,10 +256,18 @@ const ProductsList = ({
   });
 
   const handleProductSelect = (product: IProduct) => {
+    if (
+      selectionLimit !== undefined &&
+      selectedProductIds.length >= selectionLimit
+    ) {
+      return;
+    }
     setSelectedProducts((prev) => [...prev, product]);
     setSelectedProductIds((prev) => [...prev, product._id]);
   };
 
+  const selectionLimitReached =
+    selectionLimit !== undefined && selectedProductIds.length >= selectionLimit;
   const unselectedProducts = products
     .filter((p) => !selectedProductIds.includes(p._id))
     .sort(
@@ -196,11 +298,12 @@ const ProductsList = ({
 
             <SelectCategory
               selected={categoryId}
-              onSelect={
-                ((id: string) => {
-                  setCategoryId(id === categoryId ? '' : id);
-                }) as any
-              }
+              onSelect={(value) => {
+                if (typeof value !== 'string') {
+                  return;
+                }
+                setCategoryId(value === categoryId ? '' : value);
+              }}
               variant="outline"
             />
           </div>
@@ -218,6 +321,7 @@ const ProductsList = ({
                 key={product._id}
                 variant="ghost"
                 className="min-h-9 h-auto w-full justify-start font-normal whitespace-nowrap text-left"
+                disabled={selectionLimitReached}
                 onClick={() => handleProductSelect(product)}
               >
                 <div className="flex flex-1 gap-2 items-center">
@@ -261,7 +365,8 @@ const SelectedProductsList = ({
   selectedProductIds,
   setSelectedProducts,
   setSelectedProductIds,
-}: ProductsListProps) => {
+  selectedLabel,
+}: ProductsListProps & { selectedLabel: React.ReactNode }) => {
   const handleRemoveProduct = (productId: string) => {
     setSelectedProducts((prev) => prev.filter((p) => p._id !== productId));
     setSelectedProductIds((prev) => prev.filter((id) => id !== productId));
@@ -270,7 +375,9 @@ const SelectedProductsList = ({
   return (
     <ScrollArea className="h-full">
       <div className="flex flex-col gap-1 p-4">
-        <div className="px-3 mb-1 text-xs text-accent-foreground">Added</div>
+        <div className="px-3 mb-1 text-xs text-accent-foreground">
+          {selectedLabel}
+        </div>
         {selectedProductIds.map((productId) => {
           const product = selectedProducts.find((p) => p._id === productId);
           return (
