+++ frontend/plugins/sales_ui/src/modules/deals/cards/components/detail/product/constants.ts
@@ -1,0 +1,3 @@
+export const DEAL_PRODUCT_TOAST_OPTIONS = {
+  dismissOnPointerEnter: true,
+} as const;
