+++ frontend/plugins/sales_ui/src/modules/deals/cards/components/detail/product/hooks/mutations/updateDealProductsCache.ts
