+++ frontend/plugins/sales_ui/src/modules/deals/cards/components/detail/product/hooks/payment/useCheckOwnerScore.ts
@@ -1,5 +1,5 @@
 import { QueryHookOptions, useQuery } from '@apollo/client';
-import { CHECK_OWNER_SCORE } from '../graphql/queries/ScoreCampaignQueries';
+import { CHECK_OWNER_SCORE } from '../../graphql/queries/ScoreCampaignQueries';
 
 export const useCheckOwnerScore = (options: QueryHookOptions) => {
   const { data, refetch, loading } = useQuery(CHECK_OWNER_SCORE, {
