--- frontend/plugins/sales_ui/src/modules/deals/cards/components/detail/product/hooks/useProducts.tsx
@@ -1,81 +1,0 @@
-'use client';
-
-import {
-  EnumCursorDirection,
-  IRecordTableCursorPageInfo,
-  mergeCursorData,
-  useRecordTableCursor,
-  validateFetchMore,
-} from 'erxes-ui';
-import { QueryHookOptions, useQuery } from '@apollo/client';
-
-import { GET_PRODUCTS } from '../graphql/queries/ProductQueries';
-import { IProduct } from 'ui-modules';
-
-export const PRODUCTS_PER_PAGE = 30;
-
-export const useProducts = (options?: QueryHookOptions) => {
-  const { cursor } = useRecordTableCursor({
-    sessionKey: 'products-cursor',
-  });
-  const { data, loading, fetchMore } = useQuery<{
-    productsMain: {
-      list: IProduct[];
-      totalCount: number;
-      pageInfo: IRecordTableCursorPageInfo;
-    };
-  }>(GET_PRODUCTS, {
-    ...options,
-    variables: {
-      limit: PRODUCTS_PER_PAGE,
-      cursor,
-      ...options?.variables,
-    },
-  });
-
-  const { list: productsMain, totalCount, pageInfo } = data?.productsMain || {};
-
-  const handleFetchMore = ({
-    direction,
-  }: {
-    direction: EnumCursorDirection;
-  }) => {
-    if (
-      !validateFetchMore({
-        direction,
-        pageInfo,
-      })
-    ) {
-      return;
-    }
-
-    fetchMore({
-      variables: {
-        cursor:
-          direction === EnumCursorDirection.FORWARD
-            ? pageInfo?.endCursor
-            : pageInfo?.startCursor,
-        limit: PRODUCTS_PER_PAGE,
-        direction,
-      },
-      updateQuery: (prev, { fetchMoreResult }) => {
-        if (!fetchMoreResult) return prev;
-        return Object.assign({}, prev, {
-          productsMain: mergeCursorData({
-            direction,
-            fetchMoreResult: fetchMoreResult.productsMain,
-            prevResult: prev.productsMain,
-          }),
-        });
-      },
-    });
-  };
-
-  return {
-    loading,
-    productsMain,
-    totalCount,
-    handleFetchMore,
-    pageInfo,
-  };
-};
