+++ frontend/plugins/sales_ui/src/modules/deals/cards/hooks/useDeals.tsx
@@ -1,679 +1,9 @@
-import {
-  ADD_DEALS,
-  DEALS_ARCHIVE,
-  DEALS_CHANGE,
-  DEALS_COPY,
-  DEALS_WATCH,
-  EDIT_DEALS,
-  REMOVE_DEALS,
-} from '@/deals/graphql/mutations/DealsMutations';
-import {
-  EnumCursorDirection,
-  ICursorListResponse,
-  mergeCursorData,
-  toast,
-  useQueryState,
-  validateFetchMore,
-  isUndefinedOrNull,
-} from 'erxes-ui';
-import {
-  GET_DEALS,
-  GET_DEAL_DETAIL,
-} from '@/deals/graphql/queries/DealsQueries';
-import {
-  MutationHookOptions,
-  QueryHookOptions,
-  useMutation,
-  useQuery,
-} from '@apollo/client';
-import { useAtom, useAtomValue, useSetAtom } from 'jotai';
-import { useEffect, useMemo } from 'react';
-import { dealTotalCountAtom } from '@/deals/states/dealsTotalCountState';
-import { dealsViewAtom } from '@/deals/states/dealsViewState';
-import {
-  DealsBoardState,
-  useAllDealsMap,
-  useDealsBoard,
-} from '@/deals/states/dealsBoardState';
-import { useTranslation } from 'react-i18next';
-
-import { DEAL_LIST_CHANGED } from '@/deals/graphql/subscriptions/dealListChange';
-import { DEAL_CHANGED } from '@/deals/graphql/subscriptions/dealChanged';
-import { IDeal } from '@/deals/types/deals';
-import { PRODUCTS_DATA_CHANGED } from '@/deals/graphql/subscriptions/productsSubscriptions';
-import { currentUserState } from 'ui-modules';
-import { dealCreateDefaultValuesState } from '@/deals/states/dealCreateSheetState';
-import { dealDetailSheetState } from '@/deals/states/dealDetailSheetState';
-
-interface IDealChanged {
-  salesDealListChanged: {
-    action: string;
-    deal: IDeal;
-  };
-}
-
-interface ISalesProductsDataChangedPayload {
-  salesProductsDataChanged: {
-    _id: string;
-    processId: string;
-    action: string;
-    data: any;
-  };
-}
-
-interface ISalesDealChangedPayload {
-  salesDealChanged: {
-    action: string;
-    deal?: IDeal;
-  };
-}
-
-export const useDeals = (
-  options?: QueryHookOptions<ICursorListResponse<IDeal>>,
-  pipelineId?: string,
-) => {
-  const { t } = useTranslation('sales');
-  const { data, loading, fetchMore, subscribeToMore } = useQuery<
-    ICursorListResponse<IDeal>
-  >(GET_DEALS, {
-    ...options,
-    variables: { ...options?.variables },
-    skip: options?.skip,
-    fetchPolicy: 'cache-and-network',
-    onError: (e) => {
-      toast({
-        title: t('error'),
-        description: e.message,
-        variant: 'destructive',
-      });
-    },
-  });
-
-  const currentUser = useAtomValue(currentUserState);
-  const [qryStrPipelineId] = useQueryState('pipelineId');
-
-  const lastPipelineId = pipelineId || qryStrPipelineId || '';
-
-  const { list: deals, pageInfo, totalCount } = data?.deals || {};
-
-  const subscriptionVars = useMemo(
-    () => ({
-      pipelineId: lastPipelineId,
-      userId: currentUser?._id,
-      filter: options?.variables,
-    }),
-    [lastPipelineId, currentUser?._id, options?.variables],
-  );
-
-  useEffect(() => {
-    if (!currentUser?._id) return;
-
-    const unsubscribe = subscribeToMore<IDealChanged>({
-      document: DEAL_LIST_CHANGED,
-      variables: subscriptionVars,
-
-      updateQuery: (prev, { subscriptionData }) => {
-        if (!prev || !subscriptionData.data) return prev;
-        if (!prev.deals?.list) return prev;
-
-        const { action, deal } = subscriptionData.data.salesDealListChanged;
-        const currentList = prev.deals.list;
-
-        let updatedList = currentList;
-
-        if (action === 'add') {
-          const exists = currentList.some(
-            (item: IDeal) => item._id === deal._id,
-          );
-          if (!exists) {
-            const merged = [...currentList, deal];
-            merged.sort((a, b) => (a.order ?? 0) - (b.order ?? 0));
-            updatedList = merged;
-          }
-        }
-
-        if (action === 'edit') {
-          updatedList = currentList.map((item: IDeal) =>
-            item._id === deal._id ? { ...item, ...deal } : item,
-          );
-          updatedList.sort(
-            (a: IDeal, b: IDeal) => (a.order ?? 0) - (b.order ?? 0),
-          );
-        }
-
-        if (action === 'remove') {
-          updatedList = currentList.filter(
-            (item: IDeal) => item._id !== deal?._id,
-          );
-        }
-        return {
-          ...prev,
-          deals: {
-            ...prev.deals,
-            list: updatedList,
-            pageInfo: prev.deals.pageInfo,
-            totalCount:
-              action === 'add'
-                ? prev.deals.totalCount + 1
-                : action === 'remove'
-                ? prev.deals.totalCount - 1
-                : prev.deals.totalCount,
-          },
-        };
-      },
-    });
-
-    return unsubscribe;
-  }, [currentUser?._id, subscribeToMore, subscriptionVars]);
-
-  const handleFetchMore = ({
-    direction,
-  }: {
-    direction: EnumCursorDirection;
-  }) => {
-    if (!validateFetchMore({ direction, pageInfo })) {
-      return;
-    }
-
-    fetchMore({
-      variables: {
-        ...options?.variables,
-        cursor:
-          direction === EnumCursorDirection.FORWARD
-            ? pageInfo?.endCursor
-            : pageInfo?.startCursor,
-        limit: 20,
-        direction,
-      },
-      updateQuery: (prev, { fetchMoreResult }) => {
-        if (!fetchMoreResult) return prev;
-        return Object.assign({}, prev, {
-          deals: mergeCursorData({
-            direction,
-            fetchMoreResult: fetchMoreResult.deals,
-            prevResult: prev.deals,
-          }),
-        });
-      },
-    });
-  };
-
-  const view = useAtomValue(dealsViewAtom);
-  const setTotalCount = useSetAtom(dealTotalCountAtom);
-
-  useEffect(() => {
-    if (view === 'list') {
-      if (loading) {
-        setTotalCount(null);
-      } else {
-        const finalCount = isUndefinedOrNull(totalCount)
-          ? deals?.length || 0
-          : totalCount;
-        setTotalCount(finalCount);
-      }
-    }
-  }, [view, totalCount, loading, deals?.length, setTotalCount]);
-
-  useEffect(() => {
-    return () => {
-      if (view === 'list') {
-        setTotalCount(null);
-      }
-    };
-  }, [view, setTotalCount]);
-
-  return {
-    loading,
-    deals,
-    handleFetchMore,
-    pageInfo,
-    totalCount,
-  };
-};
-
-export const useDealDetail = (
-  options?: QueryHookOptions<{ dealDetail: IDeal }>,
-) => {
-  const [activeDealId] = useAtom(dealDetailSheetState);
-  const [salesItemId] = useQueryState('salesItemId');
-
-  const passedId = options?.variables?._id;
-  const finalId = passedId || salesItemId || activeDealId;
-
-  const { data, loading, error, subscribeToMore, refetch } = useQuery<{
-    dealDetail: IDeal;
-  }>(GET_DEAL_DETAIL, {
-    ...options,
-    variables: {
-      ...options?.variables,
-      _id: finalId,
-    },
-    skip: !finalId,
-    fetchPolicy: options?.fetchPolicy || 'cache-and-network',
-  });
-
-  useEffect(() => {
-    if (!salesItemId) return;
-
-    const unsubscribe = subscribeToMore<ISalesProductsDataChangedPayload>({
-      document: PRODUCTS_DATA_CHANGED,
-      variables: { _id: salesItemId },
-      updateQuery: (prev, { subscriptionData }) => {
-        const payload = subscriptionData?.data?.salesProductsDataChanged;
-        if (!payload) return prev;
-
-        const { processId } = payload;
-
-        if (processId === localStorage.getItem('processId')) {
-          return prev;
-        }
-
-        refetch();
-
-        return prev;
-      },
-    });
-
-    return unsubscribe;
-  }, [refetch, salesItemId, subscribeToMore]);
-
-  useEffect(() => {
-    if (!finalId) return;
-
-    const unsubscribe = subscribeToMore<ISalesDealChangedPayload>({
-      document: DEAL_CHANGED,
-      variables: { _id: finalId },
-      updateQuery: (prev, { subscriptionData }) => {
-        const changedDeal = subscriptionData?.data?.salesDealChanged?.deal;
-
-        if (!changedDeal) {
-          return prev;
-        }
-
-        const prevDeal = prev.dealDetail;
-        const isStageChanged =
-          changedDeal.stageId && changedDeal.stageId !== prevDeal?.stageId;
-
-        if (isStageChanged) {
-          refetch();
-        }
-
-        const pipelineId =
-          changedDeal.stage?.pipelineId ||
-          changedDeal.pipelineId ||
-          prevDeal?.pipelineId ||
-          prevDeal?.stage?.pipelineId;
-
-        return {
-          ...prev,
-          dealDetail: {
-            ...prevDeal,
-            ...changedDeal,
-            pipeline: changedDeal.pipeline || prevDeal?.pipeline,
-            pipelineId,
-            stage: changedDeal.stage || prevDeal?.stage,
-          },
-        };
-      },
-    });
-
-    return unsubscribe;
-  }, [finalId, refetch, subscribeToMore]);
-
-  const deal = data?.dealDetail;
-
-  return { deal, loading: loading && !deal, error, refetch };
-};
-
-export function useDealsEdit(options?: MutationHookOptions<any, any>) {
-  const { t } = useTranslation('sales');
-  const [_id] = useAtom(dealDetailSheetState);
-  const [salesItemId] = useQueryState('salesItemId');
-
-  const [editDeals, { loading, error }] = useMutation(EDIT_DEALS, {
-    ...options,
-    variables: {
-      ...options?.variables,
-      _id,
-    },
-    refetchQueries:
-      salesItemId || _id
-        ? [
-            {
-              query: GET_DEAL_DETAIL,
-              variables: { ...options?.variables, _id: salesItemId || _id },
-            },
-          ]
-        : [],
-    awaitRefetchQueries: true,
-    onCompleted: (...args) => {
-      toast({
-        title: t('deal-updated'),
-        variant: 'success',
-      });
-      options?.onCompleted?.(...args);
-    },
-    onError: (err) => {
-      toast({
-        title: t('error'),
-        description: err.message || t('update-failed'),
-        variant: 'destructive',
-      });
-    },
-  });
-
-  return {
-    editDeals,
-    loading,
-    error,
-  };
-}
-
-export function useDealsAdd(options?: MutationHookOptions<any, any>) {
-  const { t } = useTranslation('sales');
-  const [_id] = useAtom(dealDetailSheetState);
-  const [defaultValues] = useAtom(dealCreateDefaultValuesState);
-
-  const [addDeals, { loading, error }] = useMutation(ADD_DEALS, {
-    ...options,
-    variables: {
-      ...options?.variables,
-      _id,
-      aboveItemId: '',
-      stageId: defaultValues?.stageId,
-    },
-    awaitRefetchQueries: true,
-    onCompleted: (...args) => {
-      toast({
-        title: t('deal-added'),
-        variant: 'default',
-      });
-      options?.onCompleted?.(...args);
-    },
-    onError: (err) => {
-      toast({
-        title: t('error'),
-        description: err.message || t('update-failed'),
-        variant: 'destructive',
-      });
-    },
-  });
-
-  return {
-    addDeals,
-    loading,
-    error,
-  };
-}
-
-export function useDealsRemove(options?: MutationHookOptions<any, any>) {
-  const { t } = useTranslation('sales');
-  const [_id] = useAtom(dealDetailSheetState);
-
-  const [removeDeals, { loading, error }] = useMutation(REMOVE_DEALS, {
-    ...options,
-    variables: {
-      ...options?.variables,
-      _id,
-    },
-    awaitRefetchQueries: true,
-    onCompleted: (...args) => {
-      toast({
-        title: t('deal-removed'),
-        variant: 'default',
-      });
-      options?.onCompleted?.(...args);
-    },
-    onError: (err) => {
-      toast({
-        title: t('error'),
-        description: err.message || t('update-failed'),
-        variant: 'destructive',
-      });
-    },
-  });
-
-  return {
-    removeDeals,
-    loading,
-    error,
-  };
-}
-
-export function useDealsChange(options?: MutationHookOptions<any, any>) {
-  const { t } = useTranslation('sales');
-  const [changeDeals, { loading, error }] = useMutation(DEALS_CHANGE, {
-    ...options,
-    variables: {
-      ...options?.variables,
-    },
-    onCompleted: (...args) => {
-      toast({
-        title: t('deal-order-updated'),
-        variant: 'default',
-      });
-      options?.onCompleted?.(...args);
-    },
-    onError: (err) => {
-      toast({
-        title: t('error'),
-        description: err.message || t('update-failed'),
-        variant: 'destructive',
-      });
-    },
-  });
-
-  return {
-    changeDeals,
-    loading,
-    error,
-  };
-}
-
-export function useMoveDealStage(options?: MutationHookOptions<any, any>) {
-  const { changeDeals, loading, error } = useDealsChange(options);
-  const [boardState, setBoardState] = useDealsBoard();
-  const [allDealsMap, setAllDealsMap] = useAllDealsMap();
-
-  const moveDealStage = ({
-    deal,
-    stageId,
-    aboveItemId,
-    onCompleted,
-  }: {
-    deal: IDeal;
-    stageId: string | string[];
-    aboveItemId?: string;
-    onCompleted?: () => void;
-  }) => {
-    const destinationStageId = Array.isArray(stageId) ? stageId[0] : stageId;
-    const nextAboveItemId = aboveItemId ?? '';
-
-    if (!destinationStageId || destinationStageId === deal.stageId) {
-      onCompleted?.();
-      return Promise.resolve(null);
-    }
-
-    if (boardState) {
-      setBoardState((prev: DealsBoardState | null) => {
-        if (!prev) return prev;
-
-        const nextColumnItems = Object.fromEntries(
-          Object.entries(prev.columnItems).map(([columnId, itemIds]) => [
-            columnId,
-            (itemIds ?? []).filter((itemId: string) => itemId !== deal._id),
-          ]),
-        );
-
-        const destinationItems = nextColumnItems[destinationStageId] || [];
-        const insertIndex = nextAboveItemId
-          ? destinationItems.indexOf(nextAboveItemId) + 1
-          : 0;
-        const normalizedIndex = insertIndex > 0 ? insertIndex : 0;
-
-        destinationItems.splice(normalizedIndex, 0, deal._id);
-
-        return {
-          ...prev,
-          items: {
-            ...prev.items,
-            [deal._id]: {
-              ...(prev.items[deal._id] || deal),
-              ...deal,
-              stageId: destinationStageId,
-              columnId: destinationStageId,
-            },
-          },
-          columnItems: {
-            ...nextColumnItems,
-            [destinationStageId]: destinationItems,
-          },
-        };
-      });
-
-      setAllDealsMap((prev: Record<string, IDeal>) => ({
-        ...prev,
-        [deal._id]: {
-          ...(prev[deal._id] || deal),
-          ...deal,
-          stageId: destinationStageId,
-        },
-      }));
-    }
-
-    const processId = Math.random().toString();
-    localStorage.setItem('processId', processId);
-
-    return changeDeals({
-      variables: {
-        itemId: deal._id,
-        destinationStageId,
-        sourceStageId: deal.stageId,
-        aboveItemId: nextAboveItemId,
-        processId,
-      },
-    })
-      .then((result) => {
-        onCompleted?.();
-        return result;
-      })
-      .catch((mutationError) => {
-        if (boardState) {
-          setBoardState(boardState);
-          setAllDealsMap(allDealsMap);
-        }
-
-        throw mutationError;
-      });
-  };
-
-  return {
-    moveDealStage,
-    loading,
-    error,
-  };
-}
-
-export function useDealsArchive(options?: MutationHookOptions<any, any>) {
-  const { t } = useTranslation('sales');
-  const [archiveDealsBase, { loading, error }] = useMutation(DEALS_ARCHIVE, {
-    ...options,
-    variables: {
-      ...options?.variables,
-    },
-    awaitRefetchQueries: true,
-    onCompleted: (...args) => {
-      toast({
-        title: t('deals-archived'),
-        variant: 'default',
-      });
-      options?.onCompleted?.(...args);
-    },
-    onError: (err) => {
-      toast({
-        title: t('error'),
-        description: err.message || t('update-failed'),
-        variant: 'destructive',
-      });
-    },
-  });
-
-  const archiveDeals = (stageId: string) =>
-    archiveDealsBase({
-      variables: { stageId },
-      refetchQueries: [{ query: GET_DEALS, variables: { stageId } }],
-    });
-
-  return {
-    archiveDeals,
-    loading,
-    error,
-  };
-}
-export function useDealsCopy(options?: MutationHookOptions<any, any>) {
-  const { t } = useTranslation('sales');
-  const [_id] = useAtom(dealDetailSheetState);
-
-  const [copyDeals, { loading, error }] = useMutation(DEALS_COPY, {
-    ...options,
-    variables: {
-      ...options?.variables,
-      _id,
-    },
-    awaitRefetchQueries: true,
-    onCompleted: (...args) => {
-      toast({
-        title: t('deal-copied'),
-        variant: 'default',
-      });
-      options?.onCompleted?.(...args);
-    },
-    onError: (err) => {
-      toast({
-        title: t('error'),
-        description: err.message || t('update-failed'),
-        variant: 'destructive',
-      });
-    },
-  });
-
-  return {
-    copyDeals,
-    loading,
-    error,
-  };
-}
-
-export function useDealsWatch(options?: MutationHookOptions<any, any>) {
-  const { t } = useTranslation('sales');
-  const [_id] = useAtom(dealDetailSheetState);
-
-  const [watchDeals, { loading, error }] = useMutation(DEALS_WATCH, {
-    ...options,
-    variables: {
-      ...options?.variables,
-      _id,
-    },
-    awaitRefetchQueries: true,
-    onCompleted: (...args) => {
-      toast({
-        title: t('watch-status-updated'),
-        variant: 'default',
-      });
-      options?.onCompleted?.(...args);
-    },
-    onError: (err) => {
-      toast({
-        title: t('error'),
-        description: err.message || t('update-failed'),
-        variant: 'destructive',
-      });
-    },
-  });
-
-  return {
-    watchDeals,
-    loading,
-    error,
-  };
-}
+export { useDeals } from './deals/useDealsQuery';
+export { useDealDetail } from './deals/useDealDetail';
+export { useDealsEdit } from './deals/useDealsEdit';
+export { useDealsAdd } from './deals/useDealsAdd';
+export { useDealsRemove } from './deals/useDealsRemove';
+export { useDealsChange, useMoveDealStage } from './deals/useDealsChange';
+export { useDealsArchive } from './deals/useDealsArchive';
+export { useDealsCopy } from './deals/useDealsCopy';
+export { useDealsWatch } from './deals/useDealsWatch';
